mcp-demo/requirements.txt
# MCP 基础库
# Model Context Protocol SDK
mcp==1.5.0
# HTTP 客户端库，用于 API 调用
httpx==0.27.0
# 异步库
asyncio==3.4.3
# OpenAI
openai==1.77.0
# LangChain 相关依赖
langchain-core>=0.1.0
langchain>=0.1.0
langchain-mcp-adapters==0.0.9
langgraph>=0.5.0
langchain-openai==0.3.16
python-dotenv>=1.0.0